Revert "Revert "Okay""
This reverts commit fd8e2445f02f037f6eb958550944d1cdd18d32c9.

+++ src/servlet/RemoveContato.java
@@ -31,7 +31,7 @@ public class RemoveContato extends HttpServlet {
         PrintWriter out = response.getWriter();
 
           
-        int id = Integer.parseInt((request.getParameter("id")));
+        Long id = Long.parseLong(request.getParameter("id"));
         ContatoDao ops = new ContatoDao();
         ops.remove(id);
